pyproject.toml
[project]
name = "agentic-analytics"
version = "0.1.0"
description = "Multi-agent system for customer experience analytics"
requires-python = ">=3.11"
dependencies = [
    "chainlit>=1.3.0",
    "langgraph>=0.2.0",
    "langchain-google-vertexai>=2.0.0",
    "langchain-core>=0.3.0",
    "pandas>=2.2.0",
    "pyarrow>=15.0.0",
    "python-pptx>=1.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
    "markdown>=3.5",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.4",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.backends"
